add the with and without meta data types for drafts
