+++ libs/newsletters-data-client/src/index.ts
@@ -2,6 +2,7 @@ export * from './lib/api-response-type';
 export * from './lib/deriveNewsletterFields';
 export * from './lib/emailEmbedSchema';
 export * from './lib/legacy-newsletter-type';
+export * from './lib/meta-data-type';
 export * from './lib/newsletter-data-type';
 export * from './lib/draft-storage';
 export * from './lib/draft-to-newsletter';
