+++ libs/newsletters-data-client/src/lib/meta-data-type.ts
@@ -1,0 +1,10 @@
+import { z } from 'zod';
+
+export const metaDataSchema = z.object({
+	createdTimestamp: z.number(),
+	updatedTimestamp: z.number(),
+	createdBy: z.string(),
+	updatedBy: z.string(),
+});
+
+export type MetaData = z.infer<typeof metaDataSchema>;
